Refactor XmlRootPanelService --> XmlFoldingLoader and start work on the drawing part. Tricky stuff

+++ src/FoldingRenderer/Storage/Xml/XmlModels.cs
@@ -6,10 +6,10 @@ namespace FoldingRenderer.Storage.Xml {
     [XmlRoot("folding", Namespace = "", IsNullable = false)]
     public class Folding {
       [XmlAttribute("rootX")]
-      public double RootX { get; set; }
+      public int RootX { get; set; }
 
       [XmlAttribute("rootY")]
-      public double RootY { get; set; }
+      public int RootY { get; set; }
 
       [XmlAttribute("originalDocumentHeight")]
       public int OriginalDocumentHeight { get; set; }
